clickable choroplethmap
